@@ -1,5 +1,5 @@
 name: flutterx
-description: A new Flutter project.
+description: A Flutter project framework.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.18.6 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -31,47 +31,61 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  get: ^4.6.5
-  
-  # 环境配置
-  flutter_dotenv: ^5.1.0
-  flutter_flavor: ^3.1.1
-
-  # 日志
-  logger: ^1.1.0
-
-  # 数据存储
-  hive: ^2.2.3
-  hive_flutter: ^1.1.0
-  shared_preferences: ^2.0.8
-
-  path_provider: ^2.0.2
-
-  # 网络
-  dio: ^4.0.0
-
-  # JSON序列化
-  json_annotation: ^4.1.0
-
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
 
+  # 基础库
+  get: ^4.6.6
+  event_bus: ^2.0.0
+  logger: ^2.0.2
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+  dio: ^5.4.0
+  dio_http2_adapter: ^2.4.0
+  json_annotation: ^4.8.1
+
+  # 按环境加载配置
+  flutter_dotenv: ^5.1.0
+  flutter_flavor: ^3.1.1
+
+  # 图片库
+  flutter_svg: ^2.0.9
+  lottie: ^2.7.0
+  cached_network_image: ^3.3.0
+
+  # 屏幕适配
+  flutter_screenutil: ^5.9.0
+
+  # 平台适配
+  flutter_udid: ^2.0.1
+  path_provider: ^2.0.2
+  webview_flutter: ^4.4.2
+
+  # UI库
+  pull_to_refresh: ^2.0.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  integration_test:
+    sdk: flutter
 
-  json_serializable: ^4.1.4
+  # 测试框架
+  mockito: ^5.4.3
+  get_test: ^4.0.1
 
-  hive_generator: ^1.1.0
-  build_runner: ^2.0.6
+  # 代码生成
+  json_serializable: ^6.7.1
+  hive_generator: ^2.0.1
+  build_runner: ^2.4.7
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -88,6 +102,8 @@ flutter:
   assets:
      - .env
      - assets/mocks/
+     - assets/images/
+     - assets/web/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
