@@ -1,8 +1,7 @@
 import 'package:dio/dio.dart';
-import 'package:flutterx/app/data/model/result.dart';
-import 'package:flutterx/app/util/mock_loader.dart';
+import 'package:flutterx/app/data/model/remote/user.dart';
+import 'package:flutterx/app/core/util/mock_loader.dart';
 
-import '../../model/user.dart';
 import 'api_path.dart';
 
 abstract class UserApi {
@@ -10,21 +9,21 @@ abstract class UserApi {
 }
 
 class UserApiImpl extends UserApi {
-
   final Dio _dio;
 
   UserApiImpl(this._dio);
-  
+
   @override
   Future<User> getMe() async {
     final response = await _dio.get(ApiPath.me);
-    return Result<User>.fromJson(response.data, (json)=>User.fromJson(json as Map<String, dynamic>)).data!;
+    return User.fromJson(response.data as Map<String, dynamic>);
   }
 }
 
 class UserApiMockImpl extends UserApi {
   @override
   Future<User> getMe() {
-    return MockLoader.load('me').then((value) => Result<User>.fromJson(value, (json)=>User.fromJson(json as Map<String, dynamic>)).data!);
+    return MockLoader.load('me')
+        .then((value) => Result<User>.fromJson(value, (json) => User.fromJson(json as Map<String, dynamic>)).data!);
   }
 }
